# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pull That Up Jamie API",
    "version": "1.0.0",
    "description": "Public API for Pull That Up Jamie — a podcast search and research platform. Provides semantic search across podcast transcripts, corpus navigation for AI agents, research session management, and (upcoming) Lightning-based agent authentication.",
    "contact": {
      "name": "Pull That Up Jamie",
      "url": "https://pullthatupjamie.ai"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "https://pullthatupjamie.ai",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/research-sessions/{id}": {
      "get": {
        "tags": [
          "Research Sessions"
        ],
        "summary": "Get a research session by ID",
        "description": "Returns a specific research session with all its items and metadata. Hydrates item data from MongoDB and Pinecone as needed.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Research session ID (MongoDB ObjectId)"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-id",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "req",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Research session with items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "session": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "507f1f77bcf86cd799439011"
                            },
                            "pineconeIds": {
                              "type": "array",
                              "example": [
                                "id1",
                                "id2"
                              ],
                              "items": {
                                "type": "string"
                              }
                            },
                            "lastItemMetadata": {
                              "type": "object",
                              "properties": {}
                            },
                            "createdAt": {
                              "type": "string",
                              "example": "2026-02-13T00:00:00.000Z"
                            },
                            "updatedAt": {
                              "type": "string",
                              "example": "2026-02-13T00:00:00.000Z"
                            }
                          }
                        },
                        "items": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "session": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "507f1f77bcf86cd799439011"
                            },
                            "pineconeIds": {
                              "type": "array",
                              "example": [
                                "id1",
                                "id2"
                              ],
                              "items": {
                                "type": "string"
                              }
                            },
                            "lastItemMetadata": {
                              "type": "object",
                              "properties": {}
                            },
                            "createdAt": {
                              "type": "string",
                              "example": "2026-02-13T00:00:00.000Z"
                            },
                            "updatedAt": {
                              "type": "string",
                              "example": "2026-02-13T00:00:00.000Z"
                            }
                          }
                        },
                        "items": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Research session not found"
                    },
                    "details": {
                      "type": "string",
                      "example": "No session found for this id"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Research session not found"
                    },
                    "details": {
                      "type": "string",
                      "example": "No session found for this id"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "details": {
                      "type": "string",
                      "example": "Error fetching research session"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "details": {
                      "type": "string",
                      "example": "Error fetching research session"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/shared-research-sessions/{shareId}": {
      "get": {
        "tags": [
          "Research Sessions"
        ],
        "summary": "Get a shared research session",
        "description": "Returns metadata for a shared research session by its shareId. Intended for link unfurling and preview cards. Includes title, description, preview image, node positions, and camera configuration.",
        "parameters": [
          {
            "name": "shareId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Shared session identifier (12-character hex string)"
          }
        ],
        "responses": {
          "200": {
            "description": "Shared session metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "researchSessionId": {
                          "type": "string",
                          "example": "507f1f77bcf86cd799439011"
                        },
                        "shareId": {
                          "type": "string",
                          "example": "8d5417e36d3d"
                        },
                        "shareUrl": {
                          "type": "string",
                          "example": "https://pullthatupjamie.ai/researchSession/8d5417e36d3d"
                        },
                        "title": {
                          "type": "string",
                          "example": "My Research on Bitcoin"
                        },
                        "brandImage": {},
                        "brandColors": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        },
                        "description": {
                          "type": "string",
                          "example": "Short description or quote..."
                        },
                        "previewImageUrl": {
                          "type": "string",
                          "example": "https://.../preview.jpg"
                        },
                        "visibility": {
                          "type": "string",
                          "example": "unlisted"
                        },
                        "lastItemMetadata": {
                          "type": "object",
                          "properties": {}
                        },
                        "nodes": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        },
                        "camera": {
                          "type": "object",
                          "properties": {}
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2026-02-13T00:00:00.000Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2026-02-13T00:00:00.000Z"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "researchSessionId": {
                          "type": "string",
                          "example": "507f1f77bcf86cd799439011"
                        },
                        "shareId": {
                          "type": "string",
                          "example": "8d5417e36d3d"
                        },
                        "shareUrl": {
                          "type": "string",
                          "example": "https://pullthatupjamie.ai/researchSession/8d5417e36d3d"
                        },
                        "title": {
                          "type": "string",
                          "example": "My Research on Bitcoin"
                        },
                        "brandImage": {},
                        "brandColors": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        },
                        "description": {
                          "type": "string",
                          "example": "Short description or quote..."
                        },
                        "previewImageUrl": {
                          "type": "string",
                          "example": "https://.../preview.jpg"
                        },
                        "visibility": {
                          "type": "string",
                          "example": "unlisted"
                        },
                        "lastItemMetadata": {
                          "type": "object",
                          "properties": {}
                        },
                        "nodes": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        },
                        "camera": {
                          "type": "object",
                          "properties": {}
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2026-02-13T00:00:00.000Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2026-02-13T00:00:00.000Z"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid shareId",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Invalid shareId"
                    },
                    "details": {
                      "type": "string",
                      "example": "shareId path parameter is required"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Invalid shareId"
                    },
                    "details": {
                      "type": "string",
                      "example": "shareId path parameter is required"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Not found"
                    },
                    "details": {
                      "type": "string",
                      "example": "No shared research session found for this shareId"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Not found"
                    },
                    "details": {
                      "type": "string",
                      "example": "No shared research session found for this shareId"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "details": {
                      "type": "string",
                      "example": "Error fetching shared research session metadata"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "details": {
                      "type": "string",
                      "example": "Error fetching shared research session metadata"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentCredential": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Lightning preimage credential in format \"preimage:paymentHash\". Used for agent auth endpoints."
      },
      "BearerJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated user sessions."
      },
      "ClientId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "Anonymous client identifier for session tracking without authentication."
      }
    },
...
  }
}